Move tile highlighting code to graphics module
