+++ emergence_lib/src/graphics/mod.rs
@@ -29,6 +29,11 @@ impl Plugin for GraphicsPlugin {
                         structures::populate_structures.before(InteractionSystem::ManagePreviews),
                     ),
             )
-            .add_system_to_stage(CoreStage::PostUpdate, structures::change_structure_material);
+            .add_system_to_stage(CoreStage::PostUpdate, structures::change_structure_material)
+            .add_system(
+                terrain::display_tile_interactions
+                    .after(InteractionSystem::SelectTiles)
+                    .after(InteractionSystem::ComputeCursorPos),
+            );
     }
 }
